pi-io/requirements.txt
aiohttp==3.7.3
aiohttp-cors==0.7.0
alembic==1.4.3
aniso8601==8.1.0
async-timeout==3.0.1
attrs==20.3.0
cchardet==2.1.7
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
colorzero==1.1
Flask==1.1.2
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
gpiozero==1.5.1
idna==2.10
install==1.3.4
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-sqlalchemy==0.24.1
multidict==5.1.0
mysqlclient==2.0.3
paho-mqtt==1.5.1
pigpio==1.78
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
RPi.GPIO==0.7.0
six==1.15.0
SQLAlchemy==1.3.22
typing-extensions==3.7.4.3
urllib3==1.26.2
websockets==8.1
Werkzeug==1.0.1
yarl==1.6.3
